Refactor CleanerApp
