+++ kpops/component_handlers/helm_wrapper/utils.py
@@ -1,3 +1,4 @@
+from kpops.component_handlers.kubernetes.model import K8S_LABEL_MAX_LEN
 from kpops.component_handlers.kubernetes.utils import trim
 
 RELEASE_NAME_MAX_LEN = 53
@@ -11,3 +12,13 @@ def create_helm_release_name(name: str, suffix: str = "") -> str:
     :return: Trimmed + hashed version of the release name if it exceeds the Helm release character length otherwise the actual release name
     """
     return trim(RELEASE_NAME_MAX_LEN, name, suffix)
+
+
+def create_helm_name_override(name: str, suffix: str = "") -> str:
+    """Create Helm chart `nameOverride`.
+
+    :param name: The Helm release name to be shortened.
+    :param suffix: The release suffix to preserve
+    :return: Trimmed + hashed version of the release name if it exceeds the Kubernetes character length otherwise the actual release name
+    """
+    return trim(K8S_LABEL_MAX_LEN, name, suffix)
