+++ kpops/components/base_components/kafka_app.py
@@ -6,10 +6,7 @@ from abc import ABC
 from pydantic import AliasChoices, ConfigDict, Field
 from typing_extensions import override
 
-from kpops.component_handlers.helm_wrapper.model import (
-    HelmFlags,
-)
-from kpops.component_handlers.helm_wrapper.utils import create_helm_release_name
+from kpops.components.base_components.cleaner_app import CleanerApp
 from kpops.components.base_components.helm_app import HelmAppValues
 from kpops.components.base_components.pipeline_component import PipelineComponent
 from kpops.components.streams_bootstrap import StreamsBootstrap
@@ -51,35 +48,13 @@ class KafkaAppValues(HelmAppValues):
     )
 
 
-class KafkaAppCleaner(StreamsBootstrap, ABC):
+class KafkaAppCleaner(CleanerApp, StreamsBootstrap, ABC):
     """Helm app for resetting and cleaning a streams-bootstrap app."""
-
-    suffix: str = "-clean"
 
     @property
     @override
     def helm_chart(self) -> str:
         raise NotImplementedError
-
-    @property
-    @override
-    def full_name(self) -> str:
-        return super().full_name + self.suffix
-
-    @property
-    @override
-    def helm_release_name(self) -> str:
-        return create_helm_release_name(self.full_name, self.suffix)
-
-    @property
-    @override
-    def helm_flags(self) -> HelmFlags:
-        return HelmFlags(
-            create_namespace=self.config.create_namespace,
-            version=self.version,
-            wait=True,
-            wait_for_jobs=True,
-        )
 
     @override
     async def clean(self, dry_run: bool) -> None:
